// test-ui-generation/src/components/v2-improved-design/components/cursorFigmaMCP/CalendarModify/Calendar.module.css.d.ts
declare const styles: {
    container: string;
    monthTitle: string;
    weekDaysGrid: string;
    weekDay: string;
    daysGrid: string;
    dayContainer: string;
    day: string;
    selected: string;
    utilisation: string;
    fuite: string;
    indicator: string;
    legend: string;
    legendItem: string;
    legendDot: string;
    legendBox: string;
    legendText: string;
    utilisationColor: string;
    utilisationBorder: string;
};

export default styles;